requirements.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
coverage==7.6.12
cryptography==43.0.3
deepl==1.21.1
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dotenv-python==0.0.1
drf-yasg==1.21.8
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pygitguardian==1.19.0
Pygments==2.19.1
PyJWT==2.6.0
pyproject-api==1.9.0
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==0.21.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==12.5.1
setuptools==75.8.0
sqlparse==0.5.3
tox==4.24.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.29.2
whitenoise==6.9.0
